rename clinicalService to clinicalInfoService
