+++ containers/ecr-viewer/src/app/view-data/services/clinicalInfoService.tsx
